app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
certifi==2025.10.5
frozenlist==1.8.0
gmqtt==0.7.0
gql==4.0.0
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
multidict==6.7.0
propcache==0.4.1
PyYAML==6.0.3
sniffio==1.3.1
typing_extensions==4.15.0
websockets==15.0.1
yarl==1.22.0
